// src/entities/openapi/model/signatureTypeEnum.ts
/**
 * agridata.ch API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const SignatureTypeEnum = {
    IndividualSignature: 'INDIVIDUAL_SIGNATURE',
    CollectiveSignature: 'COLLECTIVE_SIGNATURE'
} as const;
export type SignatureTypeEnum = typeof SignatureTypeEnum[keyof typeof SignatureTypeEnum];